refactor(api,web,transactional,docs): normalize DELETE/me to 204, unswallow RSC auth errors, rename transactional utils/->lib/, add CONVENTIONS.md
- DELETE /api/v1/me now responds 204 No Content instead of returning
  { message: "Account deleted..." }. The route schema and handler
  updated; no e2e currently covers this endpoint.
- apps/web/src/lib/auth-helpers.ts drops its try/catch + console.error.
  Better Auth returns null for "no session" naturally; real failures
  propagate to Next.js error boundaries and Vercel error capture.
- packages/transactional: utils/ -> lib/ (git mv preserves history),
  barrel exports updated.
- docs/CONVENTIONS.md captures the defaults that future code reviews
  can point at.
