+++ packages/transactional/src/index.ts
@@ -5,13 +5,13 @@ export { Card } from "./components/card";
 export { Divider } from "./components/divider";
 
 // Utilities
-export { sendEmail, sendBatchEmails, previewEmail } from "./utils/send-email";
+export { sendEmail, sendBatchEmails, previewEmail } from "./lib/send-email";
 export {
   sendChangeEmailConfirmation,
   sendWelcomeEmail,
   sendPasswordResetEmail,
   sendSignUpAttemptEmail,
-} from "./utils/senders";
+} from "./lib/senders";
 
 // Theme
 export { emailTheme, tailwindConfig } from "./styles/theme";
